Trying to get popups to work on reload
