+++ app/widgets/MainNavigation.tsx
@@ -21,11 +21,6 @@ import WindowExpand = require("app/widgets/WindowExpand");
 
 import { resetUrlTimer } from 'app/url';
 
-interface ScreenPoint {
-  x: number;
-  y: number;
-}
-
 @subclass("esri.widgets.MainNavigation")
 class MainNavigation extends declared(Widget) {
   // The main map view
@@ -105,7 +100,7 @@ class MainNavigation extends declared(Widget) {
   postInitialize() {
     // Set up popup and popup event listener
     this.popup = new CustomPopup({view: this.view});
-    this.view.on('click', (event) => { this._updatePopup(event) });
+    this.view.on('click', (event) => { this.popup.openFromMouseClick(event) });
 
     this._setLoading(true);
     this.view.watch('updating', (updating) => { this._setLoading(updating) });
@@ -155,13 +150,6 @@ class MainNavigation extends declared(Widget) {
     return document.getElementById('main-navigation');
   }
 
-  // Return a feature layer by title
-  private _getLayer(name: string): FeatureLayer {
-    return (this.view.map as WebMap).layers.find((layer) => {
-      return layer.title === name;
-    }) as FeatureLayer;
-  }
-
   /*
     Set body to waiting class to display that the view is loading external
     resources. Also set the loading icon for the layers expand.
@@ -175,67 +163,6 @@ class MainNavigation extends declared(Widget) {
       document.body.classList.remove(waitingClass);
       this.layersExpand.loadingIcon = false;
     }
-  }
-
-  // Update the popup widget based on a mouse click event
-  private _updatePopup(event: any) {
-    // Reset popup variables
-    this.popup.reset();
-    this.popup.point = event.mapPoint;
-
-    const queryGeometry = this._circleAt(event.screenPoint);
-
-    [
-      this._getLayer('Sections'),
-      this._getLayer('Campus Buildings'),
-      this._getLayer('Spaces')
-    ].forEach((layer) => {
-      let query = layer.createQuery();
-      /*
-        Query features that intersect the circle around the point from
-        the click event.
-      */
-      query.geometry = queryGeometry;
-      query.spatialRelationship = 'intersects';
-
-      layer.queryFeatures(query)
-      .then((results) => {
-        if (results.features.length > 0) {
-          // Add more features to the popup
-          this.popup.features = this.popup.features.concat(results.features);
-          this.popup.visible = true;
-        }
-      }, (error) => {
-        console.error(error);
-      });
-    });
-  }
-
-  /*
-    Generate a 'circle' in latitude/longitude given a point on the screen in
-    pixels.
-  */
-  private _circleAt(screenPoint: ScreenPoint): Polygon {
-    const delta = 16;
-    const screenVertices: Array<ScreenPoint> = [
-      {x: screenPoint.x - delta, y: screenPoint.y - delta},
-      {x: screenPoint.x + delta, y: screenPoint.y - delta},
-      {x: screenPoint.x + delta, y: screenPoint.y + delta},
-      {x: screenPoint.x - delta, y: screenPoint.y + delta}
-    ]
-
-    const mapVertices = screenVertices.map((screenPoint) => {
-      const mapPoint = this.view.toMap(screenPoint);
-      /*
-        Explicitly grab lat/lon or else the polygon will try to convert back
-        to the MA spatial reference.
-      */
-      return [mapPoint.longitude, mapPoint.latitude];
-    });
-    let circle = new Polygon();
-    // Polygon ring requires the final point to be the same as the first
-    circle.addRing(mapVertices.concat([mapVertices[0]]));
-    return circle;
   }
 }
